WIP on refactoring goldens for web
