+++ packages/flutter_test/lib/src/web_golden_comparator.dart
@@ -1,35 +1,9 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 import 'dart:convert';
 import 'dart:html' as html;
-import 'dart:typed_data';
 
 import 'package:matcher/expect.dart' show fail;
 
 import 'goldens.dart';
-
-/// An unsupported [GoldenFileComparator] that exists for API compatibility.
-class LocalFileComparator extends GoldenFileComparator {
-  @override
-  Future<bool> compare(Uint8List imageBytes, Uri golden) {
-    throw UnsupportedError('LocalFileComparator is not supported on the web.');
-  }
-
-  @override
-  Future<void> update(Uri golden, Uint8List imageBytes) {
-    throw UnsupportedError('LocalFileComparator is not supported on the web.');
-  }
-}
-
-/// Returns whether [test] and [master] are pixel by pixel identical.
-///
-/// This method is not supported on the web and throws an [UnsupportedError]
-/// when called.
-Future<ComparisonResult> compareLists(List<int> test, List<int> master) async {
-  throw UnsupportedError('Golden testing is not supported on the web.');
-}
 
 /// The default [WebGoldenComparator] implementation for `flutter test`.
 ///
@@ -57,13 +31,9 @@ class DefaultWebGoldenComparator extends WebGoldenComparator {
 
   @override
   Future<bool> compare(double width, double height, Uri golden) async {
-    final String key = golden.toString();
     final html.HttpRequest request = await html.HttpRequest.request(
-      'flutter_goldens',
-      method: 'POST',
+      'flutter_screenshot',
       sendData: json.encode(<String, Object>{
-        'testUri': testUri.toString(),
-        'key': key,
         'width': width.round(),
         'height': height.round(),
       }),
